// hr-frontend/src/lib/api/generated/model/employeeProfileResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface EmployeeProfileResponse {
  profileId?: number;
  primary?: boolean;
  firstName?: string;
  otherName?: string;
  displayName?: string;
  gender?: number;
  birthDate?: string;
  nationalId?: string;
  personalEmail?: string;
  businessEmail?: string;
  phoneNumber?: string;
  mobileNumber?: string;
  countryId?: number;
  stateId?: number;
  cityId?: number;
  addressLine1?: string;
  addressLine2?: string;
  postalCode?: string;
}